- remove manual serialization to reduce liability (code is a liability)
