+++ api/hashing/v2/store_chunk_request.go
@@ -8,18 +8,8 @@ import (
 	"golang.org/x/crypto/sha3"
 )
 
-func HashStoreChunksRequest_Canonical(request *grpc.StoreChunksRequest) ([]byte, error) {
+func HashStoreChunksRequest(request *grpc.StoreChunksRequest) ([]byte, error) {
 	canonicalRequest, err := serialize.SerializeStoreChunksRequest(request)
-	if err != nil {
-		return nil, fmt.Errorf("failed to serialize store chunks request: %w", err)
-	}
-	hasher := sha3.New256()
-	_, _ = hasher.Write(canonicalRequest)
-	return hasher.Sum(nil), nil
-}
-
-func HashStoreChunksRequest_V2_Canonical(request *grpc.StoreChunksRequest) ([]byte, error) {
-	canonicalRequest, err := serialize.SerializeStoreChunksRequestV2(request)
 	if err != nil {
 		return nil, fmt.Errorf("failed to serialize store chunks request: %w", err)
 	}
